+++ Diplom/Controllers/RolesController.cs
@@ -14,9 +14,28 @@ namespace Diplom.Controllers
             _rolesService = rolesService;
         }
 
+        private async Task<StatusCodeResult> CheckRole()
+        {
+            var permissionId = int.Parse(Request.Cookies["permissions"]);
+
+            var responce = await _rolesService.GetRoleName(permissionId);
+
+            string data = responce.Data.ToLower();
+
+            if (responce.StatusCode == Domain.Enum.StatusCode.OK)
+                if (!data.Contains("admin") && !data.Contains("админ") && !data.Contains("human resources department") && !data.Contains("отдел кадров"))
+                    return Unauthorized();
+
+            return Ok();
+        }
+
         [HttpGet]
         public async Task<IActionResult> GetAllRoles()
         {
+            var result = await CheckRole();
+            if (result is UnauthorizedResult)
+                return Redirect("/");
+
             var response = await _rolesService.GetAll();
 
             if (response.StatusCode == Domain.Enum.StatusCode.OK)
@@ -28,6 +47,10 @@ namespace Diplom.Controllers
         [HttpGet]
         public async Task<IActionResult> GetRoles(int id)
         {
+            var result = await CheckRole();
+            if (result is UnauthorizedResult)
+                return Redirect("/");
+
             var response = await _rolesService.Get(id);
 
             if (response.StatusCode == Domain.Enum.StatusCode.OK)
@@ -38,6 +61,10 @@ namespace Diplom.Controllers
 
         public async Task<IActionResult> DeleteRoles(int id)
         {
+            var result = await CheckRole();
+            if (result is UnauthorizedResult)
+                return Redirect("/");
+
             var response = await _rolesService.Delete(id);
 
             if (response.StatusCode == Domain.Enum.StatusCode.OK)
@@ -49,6 +76,10 @@ namespace Diplom.Controllers
         [HttpGet]
         public async Task<IActionResult> AddOrEditRoles(int id)
         {
+            var result = await CheckRole();
+            if (result is UnauthorizedResult)
+                return Redirect("/");
+
             if (id == 0)
                 return PartialView();
 
@@ -63,6 +94,10 @@ namespace Diplom.Controllers
         [HttpPost]
         public async Task<IActionResult> AddOrEditRoles(RolesViewModel model)
         {
+            var result = await CheckRole();
+            if (result is UnauthorizedResult)
+                return Redirect("/");
+
             if (!ModelState.IsValid)
                 return View(model);
